Move test for prototype plotter
